Test the result of setup.py upload for errors.
Setup.py returns zero if the upload fails due to a 400
from the PyPI server. This can happen, for example, when
attempting to upload with a version number that has already
been uploaded.

Having the test in this script avoids the problem of forgetting
to add it in the CI server's config.