Removes CRUDError which masks the requests library HTTPError and its properties, like the request and response objects which contain information such as the status code.
CRUDError seemed to serve no purpose other than to output an error
message with very specific information.  However, all of that info
(mostly contained in the request and response objects) is logged by the
APILogger.  Given that, it makes sense to pass the requests HTTPError up
to the client rather than breaking the exception chain with CRUDError.